Implémentation de la remise pour des utilisateurs récurrents et modification de la fontion getTicket qui ne retournait pas toujours le dernier ticket

+++ src/main/java/com/parkit/parkingsystem/util/DateConvertUtil.java
@@ -41,7 +41,7 @@ public class DateConvertUtil {
     public static double getDecimalHoursFromDuration(Duration duration)
     {
         long minutes = duration.toMinutes();
-        double partOfHours = minutes / 60.0;
+        double partOfHours = (double)minutes / 60.0;
 
         return partOfHours;
     }
